Working commit of new installer system.
