+++ docker-compose.installer.yml
@@ -1,0 +1,10 @@
+services :
+  installer :
+    container_name : azuracast_installer
+    image : 'ghcr.io/azuracast/web:${AZURACAST_VERSION:-latest}'
+    volumes :
+      - './:/installer'
+    restart : 'no'
+    user : root
+    entrypoint : docker_installer
+    command : install
