+++ phpcs.xml
@@ -27,6 +27,7 @@
             warnings would make the code partly less readable
         -->
         <exclude-pattern>src/Radio/Backend/Liquidsoap/ConfigWriter.php</exclude-pattern>
+        <exclude-pattern>src/Installer/EnvFiles/*.php</exclude-pattern>
     </rule>
 
     <rule ref="SlevomatCodingStandard.Arrays.TrailingArrayComma"/>
