+++ src/Locale.php
@@ -10,6 +10,26 @@ class Locale
 {
     public const DEFAULT_LOCALE = 'en_US.UTF-8';
 
+    public const SUPPORTED_LOCALES = [
+        'en_US.UTF-8' => 'English (Default)',
+        'cs_CZ.UTF-8' => 'čeština',             // Czech
+        'de_DE.UTF-8' => 'Deutsch',             // German
+        'es_ES.UTF-8' => 'Español',             // Spanish
+        'fr_FR.UTF-8' => 'Français',            // French
+        'el_GR.UTF-8' => 'ελληνικά',            // Greek
+        'it_IT.UTF-8' => 'Italiano',            // Italian
+        'hu_HU.UTF-8' => 'magyar',              // Hungarian
+        'nl_NL.UTF-8' => 'Nederlands',          // Dutch
+        'pl_PL.UTF-8' => 'Polski',              // Polish
+        'pt_PT.UTF-8' => 'Português',           // Portuguese
+        'pt_BR.UTF-8' => 'Português do Brasil', // Brazilian Portuguese
+        'ru_RU.UTF-8' => 'Русский язык',        // Russian
+        'sv_SE.UTF-8' => 'Svenska',             // Swedish
+        'tr_TR.UTF-8' => 'Türkçe',              // Turkish
+        'zh_CN.UTF-8' => '簡化字',               // Simplified Chinese
+        'ko_KR.UTF-8' => '한국어',               // Korean (South Korean)
+    ];
+
     protected string $locale = self::DEFAULT_LOCALE;
 
     public function __construct(
@@ -25,7 +45,7 @@ class Locale
 
     protected function getValidLocale(array $possibleLocales): string
     {
-        $supportedLocales = $this->environment->getSupportedLocales();
+        $supportedLocales = self::SUPPORTED_LOCALES;
 
         foreach ($possibleLocales as $locale) {
             $locale = self::ensureLocaleEncoding($locale);
