+++ util/docker/web/scripts/docker_installer
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+cd /var/azuracast/www
+php bin/installer "$@"
